more test and cleanup repetitive configs/catalog, stop using init file for code
