+++ setup.py
@@ -16,11 +16,7 @@ setup(
     install_requires=requirements,
     entry_points="""
     [console_scripts]
-    tap-sftp=tap_sftp:main
+    tap-sftp=tap_sftp.tap_sftp:main
     """,
-    packages=["tap_sftp", "tap_sftp.singer_encodings"],
-    package_data = {
-        "schemas": ["tap_sftp/schemas/*.json"]
-    },
-    include_package_data=True,
+    packages=["tap_sftp", "tap_sftp.singer_encodings"]
 )
