+++ tests/tox_tests/test_med_00_compression.py
@@ -1,0 +1,8 @@
+from tap_sftp.singer_encodings.compression import infer
+from tests.configuration.fixtures import get_sample_file_path
+
+
+def test_infer_compression():
+    with open(get_sample_file_path('fake_file.txt.zip'), 'rb') as f:
+        out = infer(f, 'fake_file.txt.zip')
+        assert next(out).read() == b'Col1,Col2\ndata1,data2\n'
