+++ tests/tox_tests/test_med_00_discover.py
@@ -3,34 +3,9 @@ from unittest.mock import patch
 
 from singer.catalog import Catalog
 from tap_sftp.discover import discover_streams
-from tests.configuration.fixtures import get_sample_file_path
+from tests.configuration.fixtures import get_sample_file_path, get_table_spec
 
-
-@patch('tap_sftp.client.SFTPConnection.get_files_by_prefix')
-@patch('tap_sftp.client.SFTPConnection.get_file_handle')
-def test_discover_streams(patch_file_handle, patch_files):
-    # mock out the files returned from the sftp list
-    patch_files.return_value = [
-        {'filepath': '/Export/fake_file.txt', 'last_modified': datetime.now(timezone.utc)}
-    ]
-    config = {
-        'host': '',
-        'username': '',
-        'tables': [
-            {
-                'table_name': 'fake_file',
-                'search_prefix': '/Export',
-                'search_pattern': 'fake_file*',
-                'key_properties': [],
-                'delimiter': ','
-            }
-        ]
-    }
-    # mock the file handle using the local fake_file.txt
-    with open(get_sample_file_path('fake_file.txt'), 'rb') as f:
-        patch_file_handle.return_value = f
-        streams = discover_streams(config)
-    assert streams[0].get('schema') == {
+expected_schema = {
             "type": "object",
             "properties": {
                 "Col1": {
@@ -60,8 +35,23 @@ def test_discover_streams(patch_file_handle, patch_files):
             }
         }
 
-
-
+@patch('tap_sftp.client.SFTPConnection.get_files_by_prefix')
+@patch('tap_sftp.client.SFTPConnection.get_file_handle')
+def test_discover_streams(patch_file_handle, patch_files):
+    # mock out the files returned from the sftp list
+    patch_files.return_value = [
+        {'filepath': '/Export/fake_file.txt', 'last_modified': datetime.now(timezone.utc)}
+    ]
+    config = {
+        'host': '',
+        'username': '',
+        'tables': [get_table_spec()]
+    }
+    # mock the file handle using the local fake_file.txt
+    with open(get_sample_file_path('fake_file.txt'), 'rb') as f:
+        patch_file_handle.return_value = f
+        streams = discover_streams(config)
+    assert streams[0].get('schema') == expected_schema
 
 
 @patch('tap_sftp.client.SFTPConnection.get_files_by_prefix')
@@ -74,15 +64,7 @@ def test_discover_streams_decrypt(patch_file_handle, patch_files):
     config = {
         'host': '',
         'username': '',
-        'tables': [
-            {
-                'table_name': 'fake_file',
-                'search_prefix': '/Export',
-                'search_pattern': 'fake_file*',
-                'key_properties': [],
-                'delimiter': ','
-            }
-        ],
+        'tables': [get_table_spec()],
         'decryption_configs': {
             'SSM_key_name': '',
             'gnupghome': '',
@@ -93,32 +75,4 @@ def test_discover_streams_decrypt(patch_file_handle, patch_files):
     with open(get_sample_file_path('fake_file.txt'), 'rb') as f:
         patch_file_handle.return_value = f, '/temp/path/fake_file.txt'
         streams = discover_streams(config)
-    assert streams[0].get('schema') == {
-            "type": "object",
-            "properties": {
-                "Col1": {
-                    "type": [
-                        "null",
-                        "string"
-                    ]
-                },
-                "Col2": {
-                    "type": [
-                        "null",
-                        "string"
-                    ]
-                },
-                "_sdc_source_file": {
-                    "type": "string"
-                },
-                "_sdc_source_lineno": {
-                    "type": "integer"
-                },
-                "_sdc_extra": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    }
-                }
-            }
-        }
+    assert streams[0].get('schema') == expected_schema
